Don't reset UUID of ADFS users
SocialAccountAdapter used to overwrite UUID of users when they log in
via ADFS.  This would be OK for the first log-in/sign-in (because the
user doesn't have an UUID yet), but if the user has already signed in
with another account, this would change the existing UUID.  That is
problematic, because helsso clients systems might get confused about the
identity of the user.

Remove this overwriting.  Linking to correct ADFS account should happen
via the uid returned by `ADFSProvider.extract_uid`.