Overwrite django-allauth login/logout views
Use our own login/logout views rather than the ones provided by
django-allauth, because the django-allauth login view is unstyled, ugly
and has login with password functionality, which is not yet supported.

This is needed, because some django-allauth views link to its login
view.  One of those is the django-allauth's login cancelled view which
is now used if a Facebook login was cancelled by not providing e-mail
address.