Preserve state on Facebook rerequests
When Facebook login needs to rerequest the consent for e-mail address
(the feature implemented in commit 24e106c) we used to forget all state
of the login process.  This caused the login to not redirect back to
correct URL after successfully given consent for the e-mail address.

Fix this by preserving the state by passing it as a GET parameter.