OIDC: Add HelssoTokenModule
Add HelssoTokenModule for modifying the ID token generation.  Configure
django-oidc-provider to use it with OIDC_TOKEN_MODULE setting.

The configureble token module setting is my own addition to
django-oidc-provider.  Therefore also update the django-oidc-provider to
the modified version (named 0.4.4.post1+hel).  The package comes from a
personal Git branch for now, because there is a change that is not yet
ready for upstream.