User consent storing of public clients
Patch the `get` method of oidc_provider.views.AuthorizeView so that it
allows using stored user consents.

This is needed, because Django OIDC Provider doesn't allow stored User
Consent to be used if client type is public.

This hack replaces the value "public" with another value while doing the
authorize request in django_oidc and finally restores the original
value.