Add info if consent is given or not to service API endpoint
When an authenticated user fetches service data, there is a boolean field
"consent_given" which indicates whether the user has been connected to the
service. The information for this comes from UserConsent model if the service
is connected to a Client (OIDC Provider), or from AccessToken model if the
service is connected to an Application (OAuth2 Provider).