Get rid of ID token processing hook
Update django-oidc-provider to 0.6.2 which can include the claims to the
ID token with a simple setting and doesn't need ID token processing hook
for that anymore.

Set OIDC_IDTOKEN_INCLUDE_CLAIMS to True so that django-oidc-provider
will include the claims to the ID token and remove the ID token
processing hook and its setting as they are now unneeded.

Also remove functions get_userinfo_by_scopes and _get_userinfo_by_token
as they are now unused.