Add also scopes required by APIs to UserConsent objects
Previously if there was an API, which required some scope "X", that scope X
was displayed in the consent screen, but it was not added to the UserConsent
object generated as result of the authorization. Therefore it was not
possible to figure out later all the scopes a user has given consent to.

This implementation involves manipulating GET and POST params before the
authorization view, so it is a bit hacky solution, but probably still the
cleanest possible as there is no way to customize Django OIDC Provider's
authorization process.

Closes #71