Reduce amount of personal data in profile scope (#76)
Reduce amount of personal data in profile scope

Currently, the standard profile scope is requested by all client applications.

"Basic" information should definitely not contain sensitive personal data such as gender and birth date, as they are not required by any of our client applications, and we do not want them to be saved in all user databases.

Also, we do not want to ask the user for permission to give such personal data to all client applications, as current Tunnistamo implementation does; the user should know exactly what data they give permission to, and should not be asked for any data that is not needed.

Currently, the Tunnistamo db does not contain such data, but at some point in the future (e.g. suomi.fi integrations), it might come to pass that such information is available, and it should be in a strictly different scope if ever needed.

Certainly, the oidc standard lists those fields as being part of "basic" profile information, but luckily implementations such as that of OAuth (https://auth0.com/docs/scopes/current/oidc-scopes) have omitted birthdate and gender.