Add a toggle for disabling email duplicate checks
Previously Tunnistamo disallowed logging in as a user from one 3rd
party identity provider with the same email address as a previously
logged-in user from another identity provider. The user would simply
be presented with an error screen.

There is now an option and a Django settings variable to disable this
behaviour. The result will be many users inside Tunnistamo with the
same email address. This is by design, since the handling of multiple
identities will need to be custom added separately (there already are
people with multiple user accounts in the database).

Also, the criticality of the user error mentioned above is bigger with
the added idp of Suomi.fi tunnistus.