Add tokens created by OIDC provider to the Tunnistamo Session
The django_oidc_provider was forked to City of Helsinki Github because the oidc_provider didn't provide access to the tokens it created.
The changed token creation code has been added as a pull request to the original project (https://github.com/juanifioren/django-oidc-provider/pull/389) and will hopefully be merged in the future. After that we can change requirements.in to point to the original django_oidc_provider.