Get loa and amr claim values from Tunnistamo Session
Previously the loa value was saved to a Django session, but the session is not available when browser fetches tokens using an access code.

Now the values are saved into a Tunnistamo Session and read from there when updating the id token claims. The correct session can be found because we know which OIDC token (or code) corresponds to which Tunnistamo Session.

The old OptionalOpenIDScopeClaims class can be thus removed which fixes the problem that the additional claims were also shown in the userinfo.