Fix get_loa not finding the correct loa value
The id_token in extra_data is encoded token, so we can't get the claims easily. Changed the get_loa to use the already decoded id_token.

The id_token value in the backend is saved only temporary when the OIDC backend validates the incoming id_token. It's enough because we only need the value in the pipeline. The reason I used extra_data was that the loa value could be read afterwards if we had a socialauth entry. Maybe we don't need that feature.

Also changed the test backend to actually generate an encoded id_token. Previously I cut corners when creating the test backend which ended up giving false postive test results.