On social auth error redirect user to the OIDC client
After initiating an OIDC authorization, the user will select a login method and go to the social auth backend. If there is an error with the backend authentication, the user is redirected back to Tunnistamo and social auth raises an exception. The InterruptedSocialAuthMiddleware will catch the exception and always redirects the user back to the login method selector.

This patch changes the functionality so that the user is not redirected to the login method selector, but will be redirected back to the OIDC client they used. The original redirect_uri is appended with "error" and "error_description" parameters.