HP-474 Add ability to send backchannel logouts to APIs
The backchannel logout is sent to APIs when the users Tunnistamo Session ends if the API has a backchannel logout url defined.

The logout is only sent to the APIs for which the user could have fetched an API token. i.e. The APIs which scopes are mentioned in the token.