HP-918 Redirect ADFS users to ADFS logout when logging out
When the user has used a suitable ADFS backend to log in, the user is redirected straight to the ADFS log out view after they are logged out of tunnistamo.

In other cases (e.g. Google backend) the user is kept in Tunnistamo after logging out and will see the "You have been logged out of City of Helsinki services."-view.