Test original_client_id passing to HelsinkiTunnistus service
In commit 7304a7bc1ff functionality was implemented that takes the
client_id that was used by the original authorization requester, and
passes it as a parameter in the request that is made to the Helsinki
tunnistus service authorization endpoint. This functionality wasn't
fully covered by tests so this commit improves that.