Reset current active language after a test that changes it
The Django translation system works so that each thread has a current
language, and strings are translated according to that. If a test
manipulates the current language, that change remains in that thread
even after the test, if it's not explicitly reset, and that can affect
the results of the following tests.

The fixture in this commit resets the current language after a test has
finished and this new fixture is used in one test that suffered from the
above issue.