HP-1195 Change LoginMethod `name` and `short_description` to translated
Changing an existing non-translated field to a translated one requires a
multi-step process. The translated fields are first created with a
different temporary name. Then data is copied from the non-translated
fields to the translated fields. The old fields are removed and the
translated fields are renamed to have the same name as the original
fields had. All this resides in the single migration file in this
commit.

A few changes were required to an old migration file.
1) The bases needs to include TranslatableModelMixin in the migration
   where LoginMethod model is first introduced. This is required by
   Django-parler and how it works internally.
   https://github.com/django-parler/django-parler/issues/263#issuecomment-603349951
2) The non-null `name` field requires a default value so that migrating
   backwards works. The backward migration creates the non-null `name`
   column which fails unless the column also has a default value.