HP-1388 Fetch users AD groups from the Microsoft Graph
In cases where the user had over 200 groups in Azure AD the provided
claims didn't include "groups" claim at all, but included "Groups overage
claim". In which case the helusers update_ad_groups-method used in
users.pipeline.update_ad_groups pipeline stage failed because it expects the
claim to always being a list.

AD doesn't provide the groups in the response because they are worried that
the sheer size of the list would be too big to be included in headers when
using the tokens.

In our case we deliberately omit transient and non-security groups to
keep the amount of groups small.

See https://docs.microsoft.com/en-us/azure/active-directory/develop/id-tokens#groups-overage-claim
for more information about the Group Overage Claim.

Example of Group Overage Claim:

    '_claim_names': {
        'groups': 'src1'
    },
    '_claim_sources': {
        'src1': {
            'endpoint': 'https://graph.windows.net/00000000-0000-0000-0000-000000000000/users/00000000-0000-0000-0000-000000000000/getMemberObjects'
        }
    },

The endpoint mentioned in the _claim_sources doesn't seem to work with the
access tokens available when logging in. That's why we make a separate request
to Microsoft Graph to fetch the groups.

It was also decided (with Ville K) to get the AD groups from the Microsoft Graph
in all cases even when the claims include the "groups" claim to have a better
control which AD groups we provide to clients.

There are also a couple of tests for this functionality. They use the tests from
the social_core as a base.