Get Social Auth uid from the "sub" claim in HelsinkiAzureADTenantOAuth2
The base class was changed from AzureADTenantOAuth2 to AzureADV2TenantOAuth2
in the previous change. I somehow missed that the value in "preferred_username"
claim was different from the "sub" claim used in the old backend.

That change meant that Social Auth would create a new user for the users that
had previously logged in with the old backend.