Show ApiScopes' allowed apps with a better widget in admin UI
The default widget for ManyToMany fields is a simple list that needs to
be scrolled through and selections must be made with ctrl key pressed.
Now the widget is two lists next to each other, the unselected on the
left side and selected on the right. There's also a filtering option to
limit the visible entries.