HP-1699 Enable distinguishing first authorize call from the rest
During an OIDC authentication process the `authorize` endpoint may be
called many times. First when an (unauthenticated) user arrives for
authentication, it's the entrypoint for the whole authentication
process. Then a second time when authentication has been done, for
example the user returns from an IdP.

When user is redirected from the `authorize` endpoint to the login
screen, add a `first_authz=false` parameter to the "next" URL (which
points back to the `authorize` endpoint) to indicate that it's not the
first call.

Needs a new version of the forked django-oidc-provider which enables
overwriting the "redirect to login" HTTP response in the
AuthorizeEndpoint class. Fixed the django-oidc-provider version with a
commit hash instead of using a possibly moving branch name.