HP-1699 Reauthenticate with backends that are configured to do so
Social login backends can be configured to perform reauthentication.
Reauthentication happens when a user that has already previously
authenticated (i.e. has a Django session) arrives again to the
`/authorize` endpoint (with OIDC). Instead of directly returning the
user back to the service where they came from, the user is redirected to
the login endpoint, the same way as an unauthenticated user would be.