Allow passing login methods to start_oidc_authorize test function
The login methods will be configured for the oidc_client. One login
method for the passed-in backend will be added if none given in
parameters.

Previously the login methods were not configured at all. But the
functionality relied on Tunnistamo feature where if a client doesn't
have login methods set, the "global" login methods are used.