HP-1732 Add ability to filter login methods with an "idp_hint"-parameter
The parameter filters only the login methods that are used when the
login is initiated by the OIDC or OAuth2 flows.

Unknown provider ids are effectively skipped and all clients' allowed
login methods are shown if no provider ids match.

The "idp_hint"-parameter is relayed to the social auth begin-view so
that the parameter can be kept intact if the social auth redirects the
user back to the login view.