HP-1732 Keep idp_hint parameter if social auth redirects back to login
The idp_hint-parameter is saved to the session by social auth before the
user is redirected to the idp. Achieved by adding "idp_hint" to the
SOCIAL_AUTH_FIELDS_STORED_IN_SESSION-setting.

The idp_hint parameter is read from the session and appended to the
login url if InterruptedSocialAuthMiddleware redirects the user back to
the login view.