HP-1732 Consider idp_hint parameter in compatible login method check
When a previously authenticated user authenticates using another OIDC
client, it's checked that the login method that the user used before, is
good also for this new client. If not, the user gets logged out first.

Now the login methods of the new OIDC client are first filtered with the
idp_hint parameter, and only after that the check is made.