HP-1934 Prevent users from adding additional social auths
When the user was logged in with a backend that is in the
ALWAYS_REAUTHENTICATE_BACKENDS setting, there could be a case where
another log in with the same backend but with a different user would add
a second social auth to the user.

It happened if the user was still logged in to Tunnistamo but logged
out of the backend.