HP-1915 Fix server error when ADFS doesn't return unique_name
clean_attributes function tried to call split function on None.

That was because if "unique_name" was not in attrs_in the attribute
mapping would set attrs['last_first_name'] to None and the following
code only checked that the key exists in the dict before using
attrs['last_first_name'] as a string.