HP-1953 Fix bug where one AD group would be added as one char groups
If the AD returns "group" claim as a string instead of a list of strings
the update_ad_groups function would add every char in the group name as
a separate ADGroup.

Fixed by converting the string to a one element list of the string
before calling user.update_ad_groups.