Add Github action for building and publishing an image
The action triggers on pushes to the `develop` branch. It builds a
container image and pushes it to ghcr.io. Based largely on the example
from Github's documentation [1].

There used to be actions that, among other things, also built and
published a similar kind of image. But those were removed in commit
69b20c86.

[1] https://docs.github.com/en/actions/publishing-packages/publishing-docker-images#publishing-images-to-github-packages