feat: bump to django 4.2
There were significant changes in logout actions so those were the
tricky parts to handle.

- Other dependencies are bumped also where applicable.
- bump psycopg to v3
- Bump development depencies
- Bump uwsgi
- fixes migration for fresh environments
- Bump local docker postgres to 13.4 which is used in production

Data migration users.0022 uses a signal to set the data correctly.
To make the migration pass an old version of the model model is
given signal function, which makes the migration pass for any fresh
tunnistamo environment.

Refs HP-2441